[ttxn] remove TTXNWriter; set default INDENT to two spaces; update TTXNFont._tableToXML for the latest TTFont
TTXNWriter doesn't do anything more than the fontTools XMLWriter, apart from using a different three-space indentation. This looks a bit weird to me, so I made it multiple of two (as it's more often the case).
Besides, by reusing the same XMLWriter instance instead of create a new TTXNWriter, we avoid writing again the <?xml...> header in the middle of the document, which is illegal (and ugly).