Fix for crash when using a large tolerance: added check for empty glyph in restore_contour_order().
Note: The problem was that, when using a large tolerance, e.g. 50, a glyph like the U+25CC dotted circle could become an empty glyph after the little circles collapsed to colinear lines which were then removed, and restore_contour_order() assumed there was at least one contour, which resulted in a crash.