[makeotfexe] Fix issue #242, Unnecessary truncation of the Format 4 The issue was that ttx was able to compile a specific cmap specification, and makeotfexe was not. makeotfexe failed because it built a slightly larger format 4 subtable than ttx, and that was enough to overflow the subtable length limit. Compared the logic and output between the fontTools cmap module, and the makeotf logic for cmap Format 4. The fontTools logic does clearly created a more compact representation of the format 4 cmap table. Modified  the logic in makeotf's cmap.c:partitionRanges() to produce the same segment partitions as  the fontTools cmap module.
Remove code for preventing idDelta overflow in cmap format4.

The makeotfexe code has always had code to prevent an idDelta value from being assigned a value too large for a short int. This happens when the difference between a GID and its code point is greater than 32K, which is common in large CJK fonts. This certainly sounds necessary, but I noted that the fontTools code does not have this, and yet produces correct results. After both testing and code inspection, I realized that the constraints on values, and the fact that the result of adding idDelta to the code point to get GID is supposed to be modulo 64K, means that logic always works correctly.