makeotfexe] Fix issue #242, Unnecessary truncation of the Format 4.
The issue was that ttx was able to compile a specific cmap specification, and makeotfexe was not. makeotfexe failed because it built a slightly larger format 4 subtable than ttx, and that was enough to overflow the subtable length limit. Compared the logic and output between the fontTools cmap module, and the makeotf logic for cmap Format 4. The fontTools logic does clearly create a more compact representation of the format 4 cmap table. Modified the logic in makeotf's cmap.c:partitionRanges() to produce the same segment partitions as the fontTools cmap module.