[makeotfexe] Remove code for preventing idDelta overflow in cmap format 4.
The makeotfexe code has always had code to prevent an idDelta value from being assigned a value too large for a short int. This happens when the difference between a GID and its code point is greater than 32K, which is common in large CJK fonts. This certainly sounds necessary, but I noted that the fontTools code does not have this, and yet produces correct results. After both testing and code inspection, I realized that the constraints on values, and the fact that the operation of adding idDelta to the code point to get GID is supposed to be modulo 64K, means that logic always works correctly.