use setuptools_scm to manage version and files to include in sdist
The version is computed using the latest git tag, plus optionally the distance in number of commits
since the last git tag, as well as the current revision hash for the commit.

The list of files to include in the sdist is also generated using git: all the files that are
included in the repository are automatically added to the source distribution.
You can exclude some files if you wish using MANIFEST.in.