Fixed tx dcf mode with a hinted CFF2 variable font
The number of regions in VarStore table must be read ahead of reading charstrings in a CFF2 variable font, because the number of regions determines the number of parameters  to a blend operator, which in turn determines the length of each hintmask/coutermask operator.

As a fix, moved a call to dcf_getvsIndices out of dcf_DumpVarStore and placed it dcf_BegFont ahead of calling initCstrs.