[spot] ] Fix crash with 'spot -t GPOS=6' AFM-style dump
The AFM dumping logic requires creating a temporary file. With the 64-bit update, I also made a change to replace the deprecated tmpnam() function. The replacement code created a file name of length 0 for the temporary file, so the file open failed.

Note that I have now replaced the tmpnam() function with tmpfile(). This used to fail under Windows Vista through Windows 8, unless the user had admin privileges. Compiling with with VS 2017 allows this to run under Windows 7 and Windows 10. I now need to update all the Windows build files to use VS 2017 to avoid this problem with Windows users. 'spot' built with VC6 will have this problem under all platforms. We need to update to the new VS 2017 anyway, to support 64 bit tools under Windows, so I'd rather to do this than add complexity to the code to support older compilers