Fixes for Tests under Windows.
In differ.py, when 'differ' walks a dir tree to make a list of files in a ufo file, Windows and Mac list the files in a different order. The lists are compared, and the diff fails before the files are even examined.

in tx_test.py, when pdf files are compared, an additional line in the pdf file needs to be ignored. This didn't cause a problem before, but as of June 1st, the date string changed enough to make this extra line change.