Fixes for Tests under Windows. Add test for when command fails.
Note that I did not handle the 'subprocess.CalledProcessError' in runner.py::main(). I notice that all the exceptions that can be raised by the argparse module are not handled, and there is no exception handling for any other exception.   I think it is simpler to let the 'subprocess.CalledProcessError' pass through and be caught and reported by pytest.