Update Build Scripts. Fixed problem with Windows BuildAll.cmd files
I had forgotten about the delayed expansion issue with Windows command files: if a variable is defined and then used inside a if (..) clause, it is not set when it is used. You must either define it outside the clause, or make sure delayed expansion is turned on, and reference the variable usimg '!" instead of '%'.

Odd that when not using delayed expansion, it usually works - it should never work.