[makeotf] Write output OTF to same directory as input font regardless of its format
Historically, when the input font is CID the output OTF is written to the current directory, whereas in other input formats it's written to the same directory as the input font. This difference is odd, onerous and prone to error.
This change makes the output OTF be written to the same directory as the input font regardless of its format. To make the output OTF be written to the current directory use option '-o .'