[tx] Fix issue 354:  array overrun when writing a CFF2 charstring with lots of hints to CFF2. (#467)
t2cCtx *h->stack.blendArgs was sized  as 6 elements. The max number of operands in a T2 charstring (other than for the blend op) is actually 96, aka T2_Max_STEMS.

In writing test for issue 354, discovered that tx was writing the CFF2 var store incorrectly when writing a CFF2: it was adding the var store offset twice during offset calculations.