[tx] fix for issue 445: crash when tx calls fatal(),
Substituted a call to exit() instead of longjmp() in all the fatal() functions.
There are no a tests, as the only reliable test case to force the crash was removed when issue 444 was fixed. Instead, I provided tests in  a branch called 'test-for-issue445', which branched from 'develop' before the fix for issue 444. The branch 'test-for-issue445' contains a test for causing the crash before the fix is applied, and a test for verifying that the correct exit code is supplied after the fix is applied.

Note that some white space changed in 'except.c' in lines other than the fixed line; this was to make all the lines in the block have the same white space.