[makeotfexe]  Allow mark-to-base statements to reference different sets of mark classes. (#499)
makeotf posts a fatal error for mark-to-base or mark-to-mark substitutions, if two statements in the same lookup do not reference the same list of mark classes. Changed GPOS.c to allow this, and to write an offset of 0 for the unused anchor offsets in the each BaseRecord.

Added test for expected GPOS.c output. This matches the output from Google's fontmake for the same test case.