[tx] Subroutinization results differ with environment: Issue 494
Add test case to test/tx_test.py in order to validate existence of bug.

The bug is that two subr's in the win output cff are swapped in index order from the Mac version. This is because of an unstable qsort done on the subroutines in the final stage of selection.