[tx] Subroutinization results differ with environment: Issue 494
The bug is that two subr's in the win output cff are swapped in index order from the Mac version. This is because of an unstable qsort done on the subroutines in the final stage of selection.

Fix bug by adding a field to the sorted items which reflects original order, and sort by that field when the other items compare equal.
Bump version of cffwrite library
Update test to verify that both Mac and Win produce same font file.