[makeotfexe] generates incorrect secondary lookups with useMarkFilteringSet flag · Issue #196
Fix bug by adding check that lookup flag and mark filter sets match, before adding a new substitution rule to the current secondary lookup table.

This is already done in for the equivalent case in GPOS.

Update expected_output file to be the correct output.

Validate that makeotf produces the expected font.