[spot] name table difference between Mac/Win spot issue 468
Add test case to spot_test.py to verify that the test case fails under Windows and passes on the Mac.

spot currently sorts the name records by offset to the string, rather than name record order. If the offsets for two different names are the same, then the qsort is unstable, and Windows sorts differently than Mac. The fix will be  to NOT sort, and report the records by name record order.