[spot] Fix name table difference between Mac/Win spot  (#543)
spot sorted the name records by offset to the string, rather than name record order. If the offsets for two different names are the same, then the qsort is unstable, and Windows sorts differently than Mac. The fix is  to NOT sort, and report the records by name record order.

The logic was originally an optimization for name table dumps other than level 3, as report of a name string could be omitted if it the name strings were sorted by offset, and the offset of the current record was the same as the previous record. However, I now think it more valuable to have one string dump per name record, and for the strings to be reported in name record order.

Fixes #468