[makeotfexe] Only first ligature caret in first ligature is used. (#556)
[makeotfexe] Fix only first ligature caret in first ligature is used

Note that the intent of the original code not only had a bug that allowed only one ligature entry to be written, but the intent was also in error. In reviewing the OpenType GDEF table spec and the feature file syntax, I see that a ligature glyph is NOT allowed to have more than one ligature glyph table. For example, 'ffi' cannot have two ligature glyph tables, each with a different format.

Fixed issue in feat.c/GDEF.c/featgram.g.

Fixed warnings from cppcheck in GDEF.c

Updated the makefile for the ANTLR grammar generator to run the clean-up tools cpplint and clang-format after generating the feat*.[ch] files.

Update the readme file to note the addition of the cpplint and clang-format commands, and how to install them. Also mentioned that a few style errors are not fixed by clang-format, and must be fixed by hand.

Changed argument name in makeotfexe_test.py to something other than a built-in name.

Updated the expected output files.

Fixes #155