[rotatefont] Failures on Linux c-tools issue #569
 Fixed by setting the buffer length to 0 after the first pass of parsing
so that the src charstring data will get reloaded again.

Revert the test case so that under Linux the expected data is produced.

If the font is a Type1 CID font, the glyph data (charstrings) must be
read when building a subset, in order to get the FD index. When a CID
charstring is read, it is decrypted in place in the src buffer. If the
total charstring data is smaller than the src buffer, then when the
glyph data is later read again when the subset is applied, the existing
charstring data is just copied from the buffer, rather than being read
from file into the buffer. However, the data is then already decrypted,
and applying decryption again produces garbage data.

This is an old bug, and can occur on Mac and Windows as well as Linux.
It happened on Linux only because on the other OS's the std buffer size
(MAXBUFF) is defined by the compiler as  1024, while on Linux it was
6728. Hence the entire set of charstrings fit in the file buffer on
Linux.

The bug also didn't happen when the same set of glyphs was selected
using the option '-g 28-52' rather than '-fd 1', because with the former
options, the glyphs can be subset using only the glyph ID's, and the
glyphs only get read in once.