[appveyor] Leverage cibuildwheel to generate and test Windows wheels
Total revamp of our build, test, report and deploy workflow.
The setup is much simpler and all seems to be working as expected.
It has, however, one side effect: the footprint of the Windows wheel grew from 2.65MB to about 4MB. The difference is in the binaries of the C-tools. After many attempts at different build and environment settings, we were unable to determine why each of the compiled tools grew by almost 300Kb. Looking at the hex data, it seems that a locale blob got added. When we build the wheel locally its size is also around 4MB.