[makeotf test] Tweak to test_readOptionFile_relpath
If the repo is cloned to a folder at the root level — this is the case in the Linux CI — the test with input string '../different_dir' will fail because the resulting font_dir_path will be absolute (i.e. start with '/') instead of relative (i.e. start with '../'). Address this by trimming the two dots off the beginning of font_dir_path