[makeotfexe] Fix return codes.
makeotfexe should return 0 with the options -u and -h.
It should return 1 if a the source font file or source feature file is not found.
It is valid to not provide any arguments: this is the same as providing a source font file of 'font.ps', and a feature file of 'features'.