[makeotfexe] Fix old bug in setting Unicode values.
If a font has no glyphs which are assigned a Unicode value, either implicitly from having ASCII range characters, or via the GOADB file, a pointer is set to reference to whatever is in memory just before the array which is supposed to contain the UVs. On the Mac, and I imagine in prior build environments, the referenced value happens to 0, when running under 64 bit Windows, this memory os non-zero, and causes a crash. The fix is simply to skip the logic that process the array of Unicode values, when there aren't any. The bug was in  in map.c::setOS_2Fields().