[makeotfexe] Make find include feature file logic be path-separator agnostic
Fix bug that was causing the tests to fail under Windows.

The issue is that the logic for finding an included feature file failed if the last path separator, before the base file name, differed from the default for the OS.
In the file test_makeotfexe.py, the paths to the feature file included a directory, which is per OS specific.

Fixed by changing the logic so that makeotfexe will work irrespective of the OS path separator used.