[makeotfexe test] Tests for GSUB and GPOS warnings
Added tests for all the GPOS.c and GSUB.c warnings other than coverage overflow. I needed to do this because I touched all of these in order to add the parent feature name to the error report, just as I did with the coverage overflow errors.

Note that I did not do this with the warnings in feat.c. This is because the warnings in feat.c are generated during parsing ,and consequently have access to and use the parent feature file name and line number, which is better pointer to the source of a problem than the parent feature name. These are usually not available when the GPOS and GSUB structures are being filled and written, and then displaying the parent feature name is more important.