[makeotf test] Tests for validating current handling of include() in FEA files
This issue touches on two problems. The features file spec says that feature include files are relative to the parent including files, but several implementations assume that relative paths are all relative to the main feature file. The second problem is that the UFO spec says that for UFO fonts, a relative include file path is always relative to the UFO font directory. In the discussion thread for this issue, it was agreed that for non-UFO fonts, the default is to resolve relative include paths relative to the main feature file, and relative to the UFO font directory for UFO fonts. Also, for legacy project support, if an include file cannot be found, makeotfexe should also look relative to including file.

makeotfexe currently supports both of the first two implementations, but does not support include paths relative to the UFO font directory. This commit adds a test case documenting current behavior.