[ufotools] Name the UFO processed layer according to the spec
According to http://unifiedfontobject.org/versions/ufo3/layercontents.plist/ the directory name (2nd string) is derived from the layer name (1st string) and prefixed with "glyphs.".
To make this change have the least impact, the name of the directory was kept while the name of the layer got changed. However, 'processedGlyphs' was changed to 'processedglyphs' to avoid the directory name to be normalized to 'processedG_lyphs'