[makeotfexe] Fix infinite loop
On my system, makeotfexe_test.py enters into, what appears to be, an
infinite loop in test_duplicate_single_sub test.

This seems to be caused by nFeatures variable in fillFeatureList()
getting a random huge value when there are no features in the file, as
spanLimit can be 0 and the code ends up trying to access
t->subtable.array[-1].index.feature.