[makeotfexe] Fix detection of offset overflow to a FeatureParam
There were two bugs. One issue is that the overflow detection happened before the final offset adjustment was done, so that it could miss a small offset overflow. The other issue, which caused the reported bug, is that the field checked was an unsigned short, a 16 bit number, and could not hold an offset value larger than 64K.

Fixes #746