[buildcff2vf] Removed all the code within the script for merging source CFF tables into a VF CFF2 table. This now all happens within the call to varLib.build(). Provided a new function for adding glyph names to the CF font. Updated expected output ttx file. This differs from the prior version because: - fontTools rounds the delta values from real to integer in the CFF2 CharStrings - fontTools does a better job of optimizing CFF2 CharStrings
I verified that the old and new VF were functionally equivalent by manual inspection of two charstrings, and by writing a test script which did a 'tx -dump -6' of the old and VF fonts for all the source font design space coordinates , and for the coordinate (0.5,-0.5). The only differences were all in the same line segment in the same glyph. The difference was because fontTools reduces to co-linear line segments to one line segment.