[buildcff2vf] Keep all OT features when subsetting
A side effect of using the fontTools subsetter module to implement glyph subsetting with the -i option is that features are removed if they are not in the subsetter default features set. Since buildcff2vf is not intended to filter out features, I fixed this by setting the subsetter module's layout_features option field to '*', which causes it to pass through all features.

Add test case.

Fixes #817