[buildcff2vf] Fix bug in making outline points compatible
When a point is a curve in one master source font and a line in another, the code will convert the line to a flat curve. However, the code in the class CompatibilityPen assumes that the point coordinates are absolute. A few months ago, I changed the parent class CFF2CharStringMergePen (in fontTools)  from using absolute to relative coordinates, and forgot about this dependency.

Also fixed the test cases that use the compatibilization option.