[tx] report error on short CFF2 charstring resolves #895
The `bug356.otf` font originally had only one byte of a two byte integer at the end of the charstring for GID 3.

The `tx` code was treating any attempt to read past the end of a CFF2 charstring as equivalent to an `endchar` in a CFF charstring, even if it was in the middle of parsing a multi-byte operator or operand. I have changed the code to now report an error if it tries to read past the end of a CFF2 charstring while in the middle of parsing an operator or an operand. I made this fix for both the `tx -cff` and `tx -dcf` paths through the code. I also made similar fixes in some code for reading dictionaries.

I created a new folder `tests/tx_data/input/bad` for known-bad fonts, and copied the bad version of `bug356.otf` to `bug895.otf` in the `bad` folder. I then hand hex-edited the `bug356.otf` font to have a single byte integer at the end of the charstring of GID 3.

I also cleaned up several issues reported by Xcode and cppcheck. (Note that I removed some unused code as part of the cleanup.)