The logic in GPOS.c::addPosRule() for contextual mark to base rules has been deeply broken for a long time - I went back to 2014.
What needs to happen in the code starting at 2068 is that the line "nextNode->lookupLabel = anon_si->label" needs to be executed when nextNode is the base glyph/ glyph class.  This is what creates causes the actual position record to be added later. The while loop before it, however, tries to advance through the chain of nodes via nexNode->nextSeq until nextNode is either None, or is part of a mark class. This loop means that there should always have been a crash at the line which assigns "nextNode->lookupLabel = anon_si->label", as the loop always ended only when nextNode is NULL. The logic at line 2057 already sets nextNode to the base glyph node, so I fixed this problem by removing the logic that looks for the next node via some set of conditions. "nextNode" should already be the base glyph node when we get here.

There was also another serious problem. The  logic in featgram for processing the textual rule marks the first glyph in the rule as the base glyph, when it may in fact be part of a backtrack sequence. This means that the logic at line 2057 finds the first glyph node of a backtrack sequence, rather than the base glyph node. I fixed this by looking for the first marked glyph (there can be only one in this context).

I have a hard time believing that this code hasn't worked in years and we never noticed, but I can't see any other conclusion.