[makeotfexe] Fix issue 1080: Bad font built when feature file has multiple languagesystems and feature parameters #1080
Updated test file to fail the build tests with in versions of makeotf with this bug, and to pass in versions without this bug, by changing from one to three language system statements, and updating the expected ttx file.

This bug  was introduced in commit https://github.com/adobe-type-tools/afdko/commit/5624e02b58390e605639c3633025cd73cd418da7, when fixing how feature parameter tables were ordered in the GSUB/GPOS tables. The bug in in otl.c after line 2035, where bug was in incrementing nFeatureParams for every feature param subtable, instead of skipping the ones that were just references. Fixed by reverting to the original code block for this. The wrong code was left over from my first pass at changing with the subtable organization, and I forgot to revert this code when I pulled that out and used a different approach.