Initialize variables before use
When building an afdko package on the openSUSE build service
many tests failed. After some investigation I compared the workflow
in parallel debugging sessions inside the build service and a
regular system (where they didn't fail). I noticed the reason
of the failing is that afdko uses some variables that were not
initialized so on systems where memory is automatically set to 0
they work fine, but this is not safe in general (and cannot
be guaranteed), so it's better to initialize the variables.